loads of stuff
I haven't committed in a long time but have been tinkering with things
here and there.  And I think that I should be on a different branch.
Oops.

The coolest thing is the addition of a frame rate counter using TrueType
textures through FreeType.  I started with a tutorial, but I couldn't
get a FreeType atlas tutorial to work, so I backed up and learned about
textures.  Nice.