Switched FrameRateCounter to VAO
Prior to this change, VAOs were used by everything except the
FrameRateCounter.  After a long time of debugging, a rendering debacle
was uncovered that was due to object VAOs not being unbound after
rendering, causing the FreeType stuff in FrameRateCounter to walk all
over the last bound VAO.  This was fixed by unbinding VAOs after
rendering, and then it was determined that the best result would be to
make the FreeType stuff use a VAO as well.