cleanup; restriction for max string length
Without a restriction on string length, the FreeType atlas would have
had to either (1) resize its vertex buffer at runtime, which in turn
means that the method would have to be non-const, or (2) force the text
being rendered to a maximum length.  I chose the latter because, for the
moment, I am only rendering 4-5 characters for frame rate.