almost done with initial version of ParticleManager
Particles are now created, have their position updated based on their
velocity, are checked to see if they have gone out of bounds, and if
they have then they are reset.  Also have the capacity to only release a
few at a time, which was the main design difficulty in figuring out what
to do with particles that had gone out of bounds.

Still need to create a ParticleRenderer.