Finished the particle renderer
That took longer than I thought and I learned some lessons along the
way, biggest of which is that I need to create a single renderer to rule
all other renders and govern (1) the glViewport and pass the perspective
matrix on to the others, and (2) clearing the color and depth buffer.  I
wasn't paying attention to (2) and found my particles not rendering
because the buffer was being overrun by the geometry renderer.